Refactor and expand unit tests for OTP and Token
